+++ common/src/main/java/net/spell_engine/api/event/Event.java
@@ -3,6 +3,8 @@ package net.spell_engine.api.event;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.function.Consumer;
+import java.util.function.Function;
+import org.jetbrains.annotations.Nullable;
 
 public final class Event<T> {
     private boolean executing = false;
@@ -25,5 +27,21 @@ public final class Event<T> {
             function.accept(handler);
         }
         executing = false;
+    }
+
+    public <R> @Nullable R invokeWithResult(Function<T, R> function) {
+        if (executing) {
+            return null;
+        }
+        executing = true;
+        R result = null;
+        for (var handler : handlers) {
+            result = function.apply(handler);
+            if (result != null) {
+                break;
+            }
+        }
+        executing = false;
+        return result;
     }
 }