+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # 1.9.2
 
 Functional changes:
-- Fix issues of Flurry skill (windup time, animation speed, cooldown proportionality)
-- Fix issues of Swift Strikes skill (windup time)
+- Fixed issues of Flurry skill (windup time, animation speed, cooldown proportionality)
+- Fixed issues of Swift Strikes skill (windup time)
+- Fixed fallback config applying to Axes
 
 # 1.9.1